MSI-851: Associated Product popup is not opened when you want to add simple product to the configurable product
