--- app/code/Magento/InventoryCatalogAdminUi/Ui/DataProvider/Product/Listing/Modifier/Quantity.php
@@ -1,50 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\InventoryCatalogAdminUi\Ui\DataProvider\Product\Listing\Modifier;
-
-use Magento\Catalog\Ui\DataProvider\Product\Form\Modifier\AbstractModifier;
-use Magento\InventoryCatalog\Model\IsSingleSourceModeInterface;
-
-/**
- * Quantity modifier on CatalogInventory Product Grid
- */
-class Quantity extends AbstractModifier
-{
-    /**
-     * @var IsSingleSourceModeInterface
-     */
-    private $isSingleSourceMode;
-
-    /**
-     * @param IsSingleSourceModeInterface $isSingleSourceMode
-     */
-    public function __construct(
-        IsSingleSourceModeInterface $isSingleSourceMode
-    ) {
-        $this->isSingleSourceMode = $isSingleSourceMode;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function modifyData(array $data)
-    {
-        return $data;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function modifyMeta(array $meta)
-    {
-        if (false === $this->isSingleSourceMode->execute()) {
-            $meta['product_columns']['children']['qty']['arguments'] = null;
-        }
-        return $meta;
-    }
-}
