+++ app/code/Magento/InventoryCatalogAdminUi/Ui/DataProvider/Product/Listing/Modifier/QuantityPerSource.php
@@ -8,8 +8,13 @@ declare(strict_types=1);
 namespace Magento\InventoryCatalogAdminUi\Ui\DataProvider\Product\Listing\Modifier;
 
 use Magento\Catalog\Ui\DataProvider\Product\Form\Modifier\AbstractModifier;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Inventory\Model\SourceItem\Command\GetSourceItemsBySkuInterface;
+use Magento\InventoryApi\Api\SourceRepositoryInterface;
 use Magento\InventoryCatalog\Model\IsSingleSourceModeInterface;
-use Magento\InventoryCatalogAdminUi\Ui\Component\Listing\Column\SourceItems;
+use Magento\InventoryConfiguration\Model\IsSourceItemsAllowedForProductTypeInterface;
+use Magento\Ui\Component\Form\Element\DataType\Text;
+use Magento\Ui\Component\Listing\Columns\Column;
 
 /**
  * Quantity Per Source modifier on CatalogInventory Product Grid
@@ -22,20 +27,36 @@ class QuantityPerSource extends AbstractModifier
     private $isSingleSourceMode;
 
     /**
-     * @var SourceItems
+     * @var IsSourceItemsAllowedForProductTypeInterface
      */
-    private $sourceItems;
+    private $isSourceItemsAllowedForProductType;
+
+    /**
+     * @var SourceRepositoryInterface
+     */
+    private $sourceRepository;
+
+    /**
+     * @var GetSourceItemsBySkuInterface
+     */
+    private $getSourceItemsBySku;
 
     /**
      * @param IsSingleSourceModeInterface $isSingleSourceMode
-     * @param SourceItems $sourceItems
+     * @param IsSourceItemsAllowedForProductTypeInterface $isSourceItemsAllowedForProductType
+     * @param SourceRepositoryInterface $sourceRepository
+     * @param GetSourceItemsBySkuInterface $getSourceItemsBySku
      */
     public function __construct(
         IsSingleSourceModeInterface $isSingleSourceMode,
-        SourceItems $sourceItems
+        IsSourceItemsAllowedForProductTypeInterface $isSourceItemsAllowedForProductType,
+        SourceRepositoryInterface $sourceRepository,
+        GetSourceItemsBySkuInterface $getSourceItemsBySku
     ) {
         $this->isSingleSourceMode = $isSingleSourceMode;
-        $this->sourceItems = $sourceItems;
+        $this->isSourceItemsAllowedForProductType = $isSourceItemsAllowedForProductType;
+        $this->sourceRepository = $sourceRepository;
+        $this->getSourceItemsBySku = $getSourceItemsBySku;
     }
 
     /**
@@ -43,7 +64,40 @@ class QuantityPerSource extends AbstractModifier
      */
     public function modifyData(array $data)
     {
-        return $this->sourceItems->prepareDataSource(['data' => $data])['data'];
+        if (0 === $data['totalRecords'] || true === $this->isSingleSourceMode->execute()) {
+            return $data;
+        }
+
+        foreach ($data['items'] as &$item) {
+            $item['quantity_per_source'] = $this->isSourceItemsAllowedForProductType->execute($item['type_id']) === true
+                    ? $this->getSourceItemsData($item['sku'])
+                    : [];
+        }
+        unset($item);
+
+        return $data;
+    }
+
+    /**
+     * @param string $sku
+     * @return array
+     * @throws NoSuchEntityException
+     */
+    private function getSourceItemsData(string $sku): array
+    {
+        $sourceItems = $this->getSourceItemsBySku->execute($sku)->getItems();
+
+        $sourceItemsData = [];
+        foreach ($sourceItems as $sourceItem) {
+            $source = $this->sourceRepository->get($sourceItem->getSourceCode());
+            $qty = (float)$sourceItem->getQuantity();
+
+            $sourceItemsData[] = [
+                'source_name' => $source->getName(),
+                'qty' => $qty,
+            ];
+        }
+        return $sourceItemsData;
     }
 
     /**
@@ -55,49 +109,38 @@ class QuantityPerSource extends AbstractModifier
             return $meta;
         }
 
-        return array_merge_recursive($meta, [
+        $meta = array_replace_recursive($meta, [
             'product_columns' => [
                 'children' => [
-                    'quantity_per_source' => [
-                        'arguments' => $this->getArguments(),
-                        'attributes' => $this->getAttributes(),
-                        'children' => [],
-                    ]
-                ]
-            ]
+                    'quantity_per_source' => $this->getQuantityPerSourceMeta(),
+                    'qty' => [
+                        'arguments' => null,
+                    ],
+                ],
+            ],
         ]);
+        return $meta;
     }
 
     /**
      * @return array
      */
-    private function getArguments(): array
+    private function getQuantityPerSourceMeta(): array
     {
         return [
-            'data' => [
-                'config' => [
-                    'dataType' => 'text',
-                    'component' => 'Magento_InventoryCatalogAdminUi/js/product/grid/cell/source-items',
-                    'componentType' => 'column',
-                    'sortable' => false,
-                    'filter' => false,
-                    'label' => __('Quantity per Source'),
-                    'sortOrder' => 76,
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * @return array
-     */
-    private function getAttributes(): array
-    {
-        return [
-            'class' => SourceItems::class,
-            'component' => 'Magento_InventoryCatalogAdminUi/js/product/grid/cell/source-items',
-            'name' => 'quantity_per_source',
-            'sortOrder' => 76
+            'arguments' => [
+                'data' => [
+                    'config' => [
+                        'sortOrder' => 76,
+                        'filter' => false,
+                        'sortable' => false,
+                        'label' => __('Quantity per Source'),
+                        'dataType' => Text::NAME,
+                        'componentType' => Column::NAME,
+                        'component' => 'Magento_InventoryCatalogAdminUi/js/product/grid/cell/quantity-per-source',
+                    ]
+                ],
+            ],
         ];
     }
 }
