+++ app/code/Magento/InventoryCatalogAdminUi/view/adminhtml/ui_component/product_form.xml
@@ -6,7 +6,7 @@
  */
 -->
 <form xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Ui:etc/ui_configuration.xsd">
-    <fieldset name="sources" sortOrder="5" component="Magento_InventoryCatalogAdminUi/js/product/form/components/source-items">
+    <fieldset name="sources" sortOrder="5" component="Magento_InventoryCatalogAdminUi/js/product/form/source-items">
         <settings>
             <label translate="true">Sources</label>
             <collapsible>true</collapsible>
