+++ app/code/Magento/InventoryCatalogAdminUi/view/adminhtml/web/js/product/form/source-items.js
