+++ app/code/Magento/InventoryCatalogAdminUi/view/adminhtml/web/js/product/grid/cell/quantity-per-source.js
